requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.8.30
dateparser==1.2.0
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.9.11
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzlocal==5.2
yarl==1.11.1
